add new string_regex file which define structs and functions related to regex
